@@ -2,8 +2,8 @@ package com.noticias_now.ui.fragment;
 
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.noticias_now.app.ApsNoticiasApp;
 import com.noticias_now.ui.activity.BaseActivity;
